# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mountain Gear Rental API",
        "description": "API documentation for Mountain Gear Rental - Hiking Equipment Rental System. This API provides endpoints for managing hiking equipment rentals, customer data, and rental transactions.",
        "termsOfService": "https://pemweb.uas/terms",
        "contact": {
            "name": "Mountain Gear Support",
            "email": "info@mountaingear.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://pemweb.uas",
            "description": "Mountain Gear Rental API Server"
        }
    ],
    "paths": {
        "/api/hiking-equipment": {
            "get": {
                "tags": [
                    "Hiking Equipment"
                ],
                "summary": "Get all hiking equipment",
                "description": "Returns a list of all available hiking equipment for rental",
                "operationId": "getHikingEquipments",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "available",
                        "in": "query",
                        "description": "Filter by availability",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Equipment retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/HikingEquipment"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 25
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Hiking Equipment"
                ],
                "summary": "Create new hiking equipment",
                "description": "Add new hiking equipment to the rental inventory",
                "operationId": "storeHikingEquipment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "category_id",
                                    "price_per_day",
                                    "stock_quantity"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mountain Backpack Pro 60L"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Professional hiking backpack with 60L capacity"
                                    },
                                    "price_per_day": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25.5
                                    },
                                    "stock_quantity": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "brand": {
                                        "type": "string",
                                        "example": "MountainPro"
                                    },
                                    "specifications": {
                                        "type": "object",
                                        "example": {
                                            "capacity": "60L",
                                            "weight": "2.5kg"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Equipment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Equipment created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/HikingEquipment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/hiking-equipment/{id}": {
            "get": {
                "tags": [
                    "Hiking Equipment"
                ],
                "summary": "Get hiking equipment by ID",
                "description": "Returns details of specific hiking equipment",
                "operationId": "getHikingEquipmentById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Equipment ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Equipment retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/HikingEquipment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Equipment not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Hiking Equipment"
                ],
                "summary": "Update hiking equipment",
                "description": "Update existing hiking equipment details",
                "operationId": "updateHikingEquipment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Equipment ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mountain Backpack Pro 60L Updated"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "price_per_day": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 30
                                    },
                                    "stock_quantity": {
                                        "type": "integer",
                                        "example": 15
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Equipment updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Equipment updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/HikingEquipment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Equipment not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Hiking Equipment"
                ],
                "summary": "Delete hiking equipment",
                "description": "Remove hiking equipment from inventory",
                "operationId": "deleteHikingEquipment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Equipment ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Equipment deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Equipment deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Equipment not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "title": "Equipment Category",
                "description": "Category model for hiking equipment",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "Category ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Category name",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Backpacks & Carriers"
                    },
                    "slug": {
                        "description": "URL-friendly category identifier",
                        "type": "string",
                        "example": "backpacks-carriers"
                    },
                    "description": {
                        "description": "Category description",
                        "type": "string",
                        "example": "High-quality backpacks and carriers for all your hiking adventures"
                    },
                    "icon": {
                        "description": "Category icon identifier",
                        "type": "string",
                        "example": "heroicon-o-briefcase"
                    },
                    "is_active": {
                        "description": "Whether category is active",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-21T10:00:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-21T10:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "HikingEquipment": {
                "title": "Hiking Equipment",
                "description": "Hiking equipment model for rental system",
                "required": [
                    "id",
                    "name",
                    "category_id",
                    "price_per_day",
                    "stock_quantity"
                ],
                "properties": {
                    "id": {
                        "description": "Equipment ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Equipment name",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Mountain Backpack Pro 60L"
                    },
                    "slug": {
                        "description": "URL-friendly equipment identifier",
                        "type": "string",
                        "example": "mountain-backpack-pro-60l"
                    },
                    "category_id": {
                        "description": "Category ID",
                        "type": "integer",
                        "example": 1
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "description": {
                        "description": "Equipment description",
                        "type": "string",
                        "example": "Professional hiking backpack with 60L capacity, perfect for multi-day trekking adventures"
                    },
                    "price_per_day": {
                        "description": "Rental price per day in IDR",
                        "type": "number",
                        "format": "float",
                        "example": 25000
                    },
                    "stock_quantity": {
                        "description": "Available stock quantity",
                        "type": "integer",
                        "example": 10
                    },
                    "available_quantity": {
                        "description": "Currently available quantity (not rented)",
                        "type": "integer",
                        "example": 8
                    },
                    "brand": {
                        "description": "Equipment brand",
                        "type": "string",
                        "maxLength": 100,
                        "example": "MountainPro"
                    },
                    "specifications": {
                        "description": "Equipment specifications in JSON format",
                        "type": "object",
                        "example": {
                            "capacity": "60L",
                            "weight": "2.5kg",
                            "material": "Ripstop Nylon"
                        }
                    },
                    "image_url": {
                        "description": "Equipment image URL",
                        "type": "string",
                        "example": "https://images.unsplash.com/photo-1551698618-1dfe5d97d256"
                    },
                    "is_featured": {
                        "description": "Whether equipment is featured",
                        "type": "boolean",
                        "example": true
                    },
                    "is_active": {
                        "description": "Whether equipment is active for rental",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-21T10:00:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-21T10:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API Key Authentication",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}